Issue 548: extract shared env resolver code into common module (#550)
