+++ docker-compose.yml
@@ -52,6 +52,7 @@ services:
     command: node server.js
     volumes:
       - ./dev/app/:/home/node/app/
+      - ./source/lib/common:/home/node/app/lib/common
       - ./docker_volumes/node_modules_dev:/home/node/app/node_modules
     ports:
       - 0.0.0.0:3020:8080
