+++ dockerfiles/demo_dev_app/Dockerfile
@@ -9,6 +9,7 @@ RUN yarn install && yarn build
 FROM darthjee/node:0.2.1 as builder
 
 COPY --chown=node:node dev/app/ /home/node/app/
+COPY --chown=node:node source/lib/common/ /home/node/app/lib/common/
 
 RUN rm -rf spec eslint.config.mjs
