+++ source/lib/common/utils/EnvResolver.js
