+++ source/lib/common/utils/env_resolver/EnvStringResolver.js
